added public trait to course
