+++ src/plugin/cursus/Resources/modules/tools/catalog/index.js
@@ -1,5 +1,5 @@
 import {declareTool} from '#/main/core/tool'
 
-import {CatalogTool} from '#/plugin/cursus/tools/catalog/components/tool'
+import {CatalogTool} from '#/plugin/cursus/tools/catalog/containers/tool'
 
 export default declareTool(CatalogTool)
